Safari-611.1.10/LayoutTests/userscripts/user-script-image-document-expected.txt

CONSOLE MESSAGE: End Script: user-script-image-document.html
CONSOLE MESSAGE: Start Script: abe.png
CONSOLE MESSAGE: End Script: abe.png
